Homepage scrolling
The homepage should scroll back to the services section if 1. coming from a services page and 2. a previous home page position is recorded. This ensures a smooth reading experience. The scrolling was moved to the home page component because the scroll was not working when used in the layout component as some components were not completely loaded and the page could not scroll all the way down. Could have possibly still done the scrolling in the layout component, using a lifecycle method that ensured that all child compoenents are loaded?

State is used to record the last homepage scroll position and to and from routes so they are accessible anywhere